requirements.txt
python-escpos>=3.0
twilio>=8.0
flask>=3.0
